requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR & PDF Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
PyPDF2==3.0.1

# AI & NLP
openai==1.3.7
sentence-transformers==2.2.2
langdetect==1.0.9
numpy==1.26.2

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1

# Date parsing
python-dateutil==2.8.2
